% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MN simulator.R
\name{MN_sim}
\alias{MN_sim}
\title{BayICE: A hierarchical Bayesian deconvolution model with stochastic search variable selection.}
\usage{
MN_sim(seed = 123, effect_size = 0.5, n.base)
}
\arguments{
\item{seed}{seed. Default=123.}

\item{effect_size}{The normalized effect size. Default=0.5.}

\item{n.base}{The gene-specific baseline expression.}
}
\value{
The expression matrix of mixed samples and reference. A vector of cell types.
}
\description{
Multinomial-based simulator
}
\examples{
data("NSCLC")
n.base <- rowMeans(mdata.u[,grep("N",colnames(mdata.u))])
n.base <- n.base[n.base>1 & n.base < 1000]
sim.data <- MN_sim(n.base=n.base)
}
\author{
An-Shun Tai \email{daansh13@gmail.com}
}
